chore: refactored each step component into separate more descriptively named components and migrated WarningModal to input ui components dir
