+++ src/components/StepComponent.tsx
@@ -1,227 +1,10 @@
-import { ChangeEvent } from "react";
-import AllPreviewRepoCards from "./repo-cards/AllPreviewRepoCards.tsx";
-import AllTrackedRepoCards from "./repo-cards/AllTrackedRepoCards.tsx";
-import ButtonCustom from "./input/ButtonCustom.tsx";
-import { Label } from "@/components/ui/label";
-import { Separator } from "./ui/separator.tsx";
-import type {
-  StepComponentProps,
-  StepOneComponentProps,
-  StepTwoComponentProps,
-  StepThreeComponentProps,
-  StepFourComponentProps,
-} from "@/models/StepComponentModels";
+import type { StepComponentProps } from "@/models/StepComponentModels";
 
 import "../App.css";
-import InputCustom from "./input/TextInputCustom.tsx";
-import { handleStepChange } from "@/utilities/setUpUtilities.ts";
-import PaginationInput from "./input/PaginationInput.tsx";
-
-const StepOneComponent = ({
-  setHasPAT,
-  setStep,
-  setActiveNumPRs,
-  setUsername,
-  setRepoDetails,
-  setNumPageResults,
-  setDisplayWarning,
-  setReposToggled,
-  currentStep,
-  repoOwner,
-  activeNumPRs,
-}: StepOneComponentProps) => {
-  const buttonStateBundle = {
-    setStepState: setStep,
-    setPAT: setHasPAT,
-    setActiveNumPRs: setActiveNumPRs,
-    setUsername: setUsername,
-    setRepoDetails: setRepoDetails,
-    setNumPageResults: setNumPageResults,
-    setDisplayWarning: setDisplayWarning,
-    setReposToggled: setReposToggled,
-    currentStep: currentStep,
-    repoOwner: repoOwner,
-    activeNumPRs: activeNumPRs,
-  };
-
-  return (
-    <>
-      <div>
-        <div
-          style={{ marginBottom: "10px" }}
-          onClick={() =>
-            handleStepChange({
-              ...buttonStateBundle,
-              stepOperation: "stepForward",
-              initialValuePAT: null,
-            })
-          }
-        >
-          <ButtonCustom type="username" />
-        </div>
-        <Separator className="my-4" />
-
-        <div
-          onClick={() =>
-            handleStepChange({
-              ...buttonStateBundle,
-              stepOperation: "stepForward",
-              initialValuePAT: "",
-            })
-          }
-        >
-          <ButtonCustom type="usernameWithPAT" />
-        </div>
-        <div className="mt-4 text-center text-sm">
-          <a
-            href="https://github.com/cyprste2717218/github-prs-chrome-extension/tree/dev#authenticated-approach"
-            target="_blank"
-          >
-            Follow the readme here
-          </a>{" "}
-          to create a PAT (classic)
-        </div>
-      </div>
-    </>
-  );
-};
-
-const StepTwoComponent = ({
-  setUsername,
-  setRepoDetails,
-  setActiveNumPRs,
-  setStep,
-  setPAT,
-  setNumPageResults,
-  setDisplayWarning,
-  setReposToggled,
-  currentStep,
-  username,
-  activeNumPRs,
-  PAT,
-}: StepTwoComponentProps) => {
-  const handleUserNameChange = (event: ChangeEvent<HTMLInputElement>) => {
-    setUsername(event?.target?.value);
-  };
-
-  const handlePATChange = (event: ChangeEvent<HTMLInputElement>) => {
-    setPAT(event?.target?.value);
-  };
-
-  return (
-    <div
-      style={{
-        display: "flex",
-        justifyContent: "center",
-        columnGap: "10px",
-        marginTop: "20px",
-        color: "#000",
-      }}
-    >
-      <div style={{ display: "flex", flexDirection: "column" }}>
-        <div>
-          <Separator className="my-4" />
-          <div style={{ margin: "5px", textAlign: "left" }}>
-            <Label htmlFor="username">Github Username/Org Name</Label>
-          </div>
-
-          <InputCustom
-            type="username"
-            username={username}
-            handleUserNameChange={handleUserNameChange}
-          />
-        </div>
-        <div style={{ marginTop: "10px" }}>
-          {PAT !== null && (
-            <>
-              <div style={{ margin: "5px", textAlign: "left" }}>
-                <Label htmlFor="githubPAT">
-                  Your Github Personal Access Token (classic)
-                </Label>
-              </div>
-
-              <InputCustom
-                type="PAT"
-                PAT={PAT}
-                handlePATChange={handlePATChange}
-              />
-            </>
-          )}
-        </div>
-        <div style={{ marginTop: "15px" }}>
-          <ButtonCustom
-            type="submit"
-            username={username}
-            patCode={PAT}
-            currentStep={currentStep}
-            repoOwner={username}
-            activeNumPRs={activeNumPRs}
-            initialValuePAT={PAT}
-            setRepoDetails={setRepoDetails}
-            setStep={setStep}
-            setPAT={setPAT}
-            setActiveNumPRs={setActiveNumPRs}
-            setUsername={setUsername}
-            setNumPageResults={setNumPageResults}
-            setDisplayWarning={setDisplayWarning}
-            setReposToggled={setReposToggled}
-          />
-        </div>
-      </div>
-    </div>
-  );
-};
-
-const StepThreeComponent = ({
-  repoDetails,
-  setActiveNumPRs,
-  setNumPageResults,
-  setRepoDetails,
-  activeNumPRs,
-  step,
-  numPageResults,
-  username,
-  patCode,
-  reposToggled,
-}: StepThreeComponentProps) => {
-  return (
-    <>
-      <AllPreviewRepoCards
-        setActiveNumPRs={setActiveNumPRs}
-        setNumPageResults={setNumPageResults}
-        activeNumPRs={activeNumPRs}
-        repoDetails={repoDetails}
-        step={step}
-        numPageResults={numPageResults}
-        setRepoDetails={setRepoDetails}
-        username={username}
-        patCode={patCode}
-        reposToggled={reposToggled}
-      />
-      <PaginationInput
-        setNumPageResults={setNumPageResults}
-        numPageResults={numPageResults}
-        setRepoDetails={setRepoDetails}
-        username={username}
-        patCode={patCode}
-      />
-    </>
-  );
-};
-
-const StepFourComponent = ({
-  activeNumPRs,
-  githubUsername,
-}: StepFourComponentProps) => {
-  return (
-    <>
-      <AllTrackedRepoCards
-        githubUsername={githubUsername}
-        activeNumPRs={activeNumPRs}
-      />
-    </>
-  );
-};
+import ChooseSetupOptComponent from "./ChooseSetupOptComponent.tsx";
+import GitDetailsEntryComponent from "./GitDetailsEntryComponent.tsx";
+import ChooseReposComponent from "./ChooseReposComponent.tsx";
+import DisplayTrackedReposComponent from "./DisplayTrackedReposComponent.tsx";
 
 const StepComponent = ({
   setUsername,
@@ -246,7 +29,7 @@ const StepComponent = ({
   switch (step) {
     case 1:
       CurrentStepUI = (
-        <StepOneComponent
+        <ChooseSetupOptComponent
           setStep={setStep}
           setHasPAT={setHasPAT}
           setActiveNumPRs={setActiveNumPRs}
@@ -264,7 +47,7 @@ const StepComponent = ({
 
     case 2:
       CurrentStepUI = (
-        <StepTwoComponent
+        <GitDetailsEntryComponent
           setUsername={setUsername}
           setRepoDetails={setRepoDetails}
           setStep={setStep}
@@ -284,7 +67,7 @@ const StepComponent = ({
 
     case 3:
       CurrentStepUI = (
-        <StepThreeComponent
+        <ChooseReposComponent
           setActiveNumPRs={setActiveNumPRs}
           setNumPageResults={setNumPageResults}
           setRepoDetails={setRepoDetails}
@@ -301,7 +84,7 @@ const StepComponent = ({
 
     case 4:
       CurrentStepUI = (
-        <StepFourComponent
+        <DisplayTrackedReposComponent
           githubUsername={username}
           activeNumPRs={activeNumPRs}
         />
