+++ tsconfig.app.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/app.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/components/stepcomponent.tsx","./src/components/warningmodal.tsx","./src/components/header/headercomponent.tsx","./src/components/header/titlecomponent.tsx","./src/components/input/buttoncustom.tsx","./src/components/input/checkboxcustom.tsx","./src/components/input/paginationinput.tsx","./src/components/input/selectallbutton.tsx","./src/components/input/textinputcustom.tsx","./src/components/repo-cards/allpreviewrepocards.tsx","./src/components/repo-cards/alltrackedrepocards.tsx","./src/components/ui/badge.tsx","./src/components/ui/button.tsx","./src/components/ui/card.tsx","./src/components/ui/checkbox.tsx","./src/components/ui/collapsible.tsx","./src/components/ui/dialog.tsx","./src/components/ui/input.tsx","./src/components/ui/label.tsx","./src/components/ui/loadingbutton.tsx","./src/components/ui/pagination.tsx","./src/components/ui/separator.tsx","./src/lib/utils.ts","./src/models/checkboxcustommodels.ts","./src/models/headercomponentmodels.ts","./src/models/inputmodels.tsx","./src/models/repocardmodels.ts","./src/models/stepcomponentmodels.ts","./src/models/textinputmodels.ts","./src/models/stephandlemodels.ts","./src/utilities/repodetailutilities.ts","./src/utilities/setuputilities.ts"],"version":"5.8.3"}
+{"root":["./src/app.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/components/choosereposcomponent.tsx","./src/components/choosesetupoptcomponent.tsx","./src/components/displaytrackedreposcomponent.tsx","./src/components/gitdetailsentrycomponent.tsx","./src/components/stepcomponent.tsx","./src/components/header/headercomponent.tsx","./src/components/header/titlecomponent.tsx","./src/components/input/buttoncustom.tsx","./src/components/input/checkboxcustom.tsx","./src/components/input/paginationinput.tsx","./src/components/input/selectallbutton.tsx","./src/components/input/textinputcustom.tsx","./src/components/input/warningmodal.tsx","./src/components/repo-cards/allpreviewrepocards.tsx","./src/components/repo-cards/alltrackedrepocards.tsx","./src/components/ui/badge.tsx","./src/components/ui/button.tsx","./src/components/ui/card.tsx","./src/components/ui/checkbox.tsx","./src/components/ui/collapsible.tsx","./src/components/ui/dialog.tsx","./src/components/ui/input.tsx","./src/components/ui/label.tsx","./src/components/ui/loadingbutton.tsx","./src/components/ui/pagination.tsx","./src/components/ui/separator.tsx","./src/lib/utils.ts","./src/models/checkboxcustommodels.ts","./src/models/headercomponentmodels.ts","./src/models/inputmodels.tsx","./src/models/repocardmodels.ts","./src/models/stepcomponentmodels.ts","./src/models/textinputmodels.ts","./src/models/stephandlemodels.ts","./src/utilities/repodetailutilities.ts","./src/utilities/setuputilities.ts"],"version":"5.8.3"}
